build base packet changed to build base layers

+++ PcapFunc/ProtocolFields/Control/Handshake.cs
@@ -4,7 +4,7 @@ namespace SRTManager.ProtocolFields.Control
 {
     public class Handshake : SRTHeader
     {
-        public Handshake(uint version, ushort encryption_field, uint intial_psn, uint type, uint source_socket_id, uint dest_socket_id, uint syn_cookie, double p_ip) : base(ControlType.HANDSHAKE, dest_socket_id)
+        public Handshake(uint version, ushort encryption_field, uint intial_psn, uint type, uint source_socket_id, uint dest_socket_id, uint syn_cookie, long p_ip) : base(ControlType.HANDSHAKE, dest_socket_id)
         {
             VERSION = version; byteFields.Add(BitConverter.GetBytes(VERSION));
             ENCRYPTION_FIELD = encryption_field; byteFields.Add(BitConverter.GetBytes(ENCRYPTION_FIELD));
@@ -14,7 +14,7 @@ namespace SRTManager.ProtocolFields.Control
             TYPE = type; byteFields.Add(BitConverter.GetBytes(TYPE));
             SOCKET_ID = source_socket_id; byteFields.Add(BitConverter.GetBytes(SOCKET_ID));
             SYN_COOKIE = syn_cookie; byteFields.Add(BitConverter.GetBytes(SYN_COOKIE));
-            PEER_IP = p_ip; byteFields.Add(BitConverter.GetBytes(Convert.ToDouble(PEER_IP)));
+            PEER_IP = p_ip; byteFields.Add(BitConverter.GetBytes(Convert.ToInt64(PEER_IP)));
         }
 
         public Handshake(byte[] data) : base(data)  // initialize SRT Control header fields
@@ -29,7 +29,7 @@ namespace SRTManager.ProtocolFields.Control
             TYPE = BitConverter.ToUInt32(data, 31);  // [31 32 33 34] (4 bytes)
             SOCKET_ID = BitConverter.ToUInt32(data, 35);  // [35 36 37 38] (4 bytes)
             SYN_COOKIE = BitConverter.ToUInt32(data, 39);  // [39 40 41 42] (4 bytes)
-            PEER_IP = BitConverter.ToUInt64(data, 43);  // [43 44 45 46 47 48 49 50] (8 bytes)
+            PEER_IP = BitConverter.ToInt64(data, 43);  // [43 44 45 46 47 48 49 50] (8 bytes)
         }
 
         public static bool IsHandshake(byte[] data)
@@ -96,7 +96,7 @@ namespace SRTManager.ProtocolFields.Control
         /// sender.The value consists of four 32-bit fields.In the case of
         /// IPv4 addresses, fields 2, 3 and 4 are filled with zeroes.
         /// </summary>
-        public double PEER_IP { get; set; }
+        public long PEER_IP { get; set; }
 
 
         public enum Extension // Extension Field
